Remove the dependency on uuid.
